Bug fix to canvas.saveLayer....
An issue came up where by the canvas.saveLayer wasn't working correctly with restoreCount. This was due to the API version being too old. I've added in a check and if the API version is above API 26 then it will use the newer version of the method otherwise use old version.

Also changed constructor to public.